refactor : Treatments 패키지 변경
