+++ src/main/java/com/example/hospitalreservation/service/treatmentfees/TreatmentsFee.java
@@ -1,6 +1,6 @@
 package com.example.hospitalreservation.service.treatmentfees;
 
-import com.example.hospitalreservation.enums.Treatments;
+import com.example.hospitalreservation.domain.Treatments;
 
 public interface TreatmentsFee extends Fee {
